chore: reorganize pages folder structure to comply with standards
