+++ src/hooks/MoneyDestination/useAddDestination.ts
@@ -3,8 +3,8 @@ import { useNavigate } from "react-router-dom";
 import { FormikProps } from "formik";
 import { IRuleDecision } from "@isettingkit/input";
 
-import { addDestinationStepsConfig } from "@pages/moneyDestination/tabs/moneyDestinationTab/AddDestination/config/assisted.config";
-import { IGeneralInformationEntry } from "@pages/moneyDestination/tabs/moneyDestinationTab/forms/GeneralInformation/types";
+import { addDestinationStepsConfig } from "@pages/moneyDestination/tabs/moneyDestinationTab/addDestination/config/assisted.config";
+import { IGeneralInformationEntry } from "@pages/moneyDestination/tabs/moneyDestinationTab/forms/generalInformation/types";
 import { IRequestSteps } from "@design/feedback/RequestProcess/types";
 
 const useAddDestination = (requestSteps: IRequestSteps[]) => {
