+++ src/hooks/creditLine/useAddCreditLines/index.ts
@@ -2,7 +2,7 @@ import { useRef, useState } from "react";
 import { FormikProps } from "formik";
 import { IRuleDecision } from "@isettingkit/input";
 
-import { IGeneralInformationEntry } from "@pages/creditLines/forms/GeneralInformation/types";
+import { IGeneralInformationEntry } from "@src/pages/creditLines/forms/generalInformation/types";
 import { ICreditProspectEntry } from "@design/forms/creditProspect/types";
 import { optionsProspectConfig } from "@design/forms/creditProspect/config/optionsProspect.config";
 import {
