+++ src/pages/moneyDestination/tabs/moneyDestinationTab/addDestination/config/assisted.config.ts
