+++ src/pages/moneyDestination/tabs/moneyDestinationTab/addDestination/config/navigation.ts
