+++ src/pages/moneyDestination/tabs/moneyDestinationTab/forms/generalInformation/index.tsx
