+++ src/pages/moneyDestination/tabs/moneyDestinationTab/forms/verificationForm/verificationBoxes/styles.ts
