+++ src/routes/selectBusinessunits.tsx
@@ -2,9 +2,9 @@ import { Route, Routes } from "react-router-dom";
 import { useContext } from "react";
 
 import { ErrorPage } from "@components/layout/ErrorPage";
-import { IBusinessUnit } from "@pages/selectBusinessUnits/outlets/BusinessUnit/types";
+import { IBusinessUnit } from "@pages/selectBusinessUnits/outlets/businessUnit/types";
 import { SelectBusinessUnits } from "@pages/selectBusinessUnits";
-import { BusinessUnits } from "@pages/selectBusinessUnits/outlets/BusinessUnit";
+import { BusinessUnits } from "@pages/selectBusinessUnits/outlets/businessUnit";
 import { CheckingCredentials } from "@pages/login/checkingCredentials";
 import { LoadingApp } from "@components/feedback/LoadingApp";
 import { NotBusinessUnit } from "@pages/errors/notBusinessUnit";
